#13 login progress
might have to rethink the modal login thing in relation to inertia. It seems like better UX, but is a bit of a pain in terms of error handling.

There might be a way to control showing a login modal from a route server side and have it play nicely with inertia, something about query params.